requirements.txt
# Core
torch==2.6.0
torchvision==0.21.0
accelerate>=0.34.0
safetensors
einops
numpy>=1.26
opencv-python
imageio>=2.34
tqdm>=4.66
decord
moviepy

# Diffusion / LoRA
diffusers>=0.32.0
transformers>=4.44.0
peft>=0.13.0

# WAN / pipelines (leave hook-points; repo code will import local modules)
omegaconf>=2.3.0
PyYAML>=6.0.1
